validation: moved required params to folder API
